test(suite): co-locate quote hook tests
