+++ packages/suite/src/hooks/wallet/trading/form/buy/useBuyQuotes.test.tsx
@@ -1,33 +1,33 @@
 import { type Resolver, useForm } from 'react-hook-form';
 
 import { act, waitFor } from '@testing-library/react';
-import { type CryptoId, type SellFiatTrade } from 'invity-api';
+import type { BuyTrade, CryptoId } from 'invity-api';
 
 import { configureMockStore, renderHookWithStoreProvider } from '@suite-common/test-utils';
 import {
-    type TradingAssetSellOption,
-    type TradingSellFormProps,
-    sellInitialState,
+    type TradingAssetOption,
+    type TradingBuyFormProps,
+    type TradingCountryOption,
+    buyInitialState,
     initialState as tradingInitialState,
 } from '@suite-common/trading';
-import { type Network, getNetwork } from '@suite-common/wallet-config';
-import { mockAccountKey } from '@suite-common/wallet-types/mocks';
+import { getNetwork } from '@suite-common/wallet-config';
 
-import { DEBOUNCE_DELAY_MS } from '../../common/useTradingQuoteRequest';
-import { useSellQuotes } from '../useSellQuotes';
+import { useBuyQuotes } from './useBuyQuotes';
+import { DEBOUNCE_DELAY_MS } from '../common/useTradingQuoteRequest';
 
-const QUOTES: SellFiatTrade[] = [
-    {
-        paymentMethod: 'bankTransfer',
-        paymentMethodName: 'Bank transfer',
-        exchange: 'provider-1',
-        rate: 2,
-    },
+const QUOTES: BuyTrade[] = [
     {
         paymentMethod: 'creditCard',
         paymentMethodName: 'Credit card',
-        exchange: 'provider-2',
+        exchange: 'provider-1',
         rate: 1,
+    },
+    {
+        paymentMethod: 'bankTransfer',
+        paymentMethodName: 'Bank transfer',
+        exchange: 'provider-2',
+        rate: 2,
     },
 ];
 
@@ -37,7 +37,6 @@ const mockHandleRequest = jest.fn((payload: unknown) => {
 
     return Object.assign(thunk, { payload });
 });
-const mockClearQuotes = jest.fn();
 
 jest.mock('@suite-common/dependency-injection', () => ({
     ...jest.requireActual('@suite-common/dependency-injection'),
@@ -49,70 +48,24 @@ jest.mock('@suite-common/trading', () => {
 
     return {
         ...actual,
-        sellThunks: {
-            ...actual.sellThunks,
+        buyThunks: {
+            ...actual.buyThunks,
             handleRequestThunk: (payload: unknown) => mockHandleRequest(payload),
-        },
-        tradingSellActions: {
-            ...actual.tradingSellActions,
-            clearQuotes: () => {
-                mockClearQuotes();
-
-                return actual.tradingSellActions.clearQuotes();
-            },
         },
     };
 });
 
-const SEND_CRYPTO_SELECT: TradingAssetSellOption = {
-    id: 'bitcoin' as CryptoId,
-    isNativeToken: true,
-    name: 'Bitcoin',
-    coingeckoId: 'bitcoin',
-    contractAddress: null,
-    symbol: 'btc',
-    displaySymbol: 'BTC',
-    networkName: 'Bitcoin',
-    networkSymbol: 'btc',
-    accountKey: mockAccountKey({ descriptor: 'descriptor123', symbol: 'btc' }),
-};
-
-const VALID_DEFAULTS: TradingSellFormProps = {
-    outputs: [
-        {
-            type: 'payment',
-            address: 'address',
-            amount: '0.0015',
-            fiat: '50',
-            currency: { value: 'usd', label: 'USD' },
-            token: null,
-            label: '',
-        },
-    ],
-    countrySelect: {
-        value: 'CZ' as const,
-        codeAlpha3: 'CZE',
-        flag: '🇨🇿',
-        name: 'Czechia',
-        label: '🇨🇿 Czechia',
-        shortLabel: '🇨🇿 CZE',
-    },
-    sendCryptoSelect: SEND_CRYPTO_SELECT,
-    amountInCrypto: true,
+const VALID_DEFAULTS: TradingBuyFormProps = {
+    fiatInput: '100',
+    cryptoInput: '',
+    currencySelect: { value: 'eur', label: 'EUR' },
+    cryptoSelect: { id: 'bitcoin' as CryptoId, networkSymbol: 'btc' } as TradingAssetOption,
+    countrySelect: { value: 'DE', label: 'Germany' } as TradingCountryOption,
+    countrySubdivisionSelect: undefined,
     paymentMethod: undefined,
     provider: undefined,
-    feePerUnit: '',
-    feeLimit: '',
-    options: ['broadcast'],
-    bitcoinLocktimeBlockHeight: '',
-    bitcoinLocktimeDatetime: '',
-    ethereumNonce: '',
-    transactionData: '',
-    destinationTag: '',
-    isCoinControlEnabled: false,
-    hasCoinControlBeenOpened: false,
-    utxoSorting: 'newestFirst',
-    selectedUtxos: [],
+    amountInCrypto: false,
+    receiveAddress: 'bc1qreceive',
 };
 
 const NO_REFETCH_WAIT_MS = DEBOUNCE_DELAY_MS + 200;
@@ -125,55 +78,45 @@ const wait = (ms: number) =>
             }),
     );
 
-const renderSellQuotes = (
-    defaultValues: TradingSellFormProps,
-    options: { resolver?: Resolver<TradingSellFormProps> } = {},
+const renderBuyQuotes = (
+    defaultValues: TradingBuyFormProps,
+    options: { resolver?: Resolver<TradingBuyFormProps> } = {},
 ) => {
     const { resolver } = options;
-    const initialProps: { currentNetwork: Network | undefined } = {
-        currentNetwork: getNetwork('btc'),
-    };
-
     const store = configureMockStore({
         preloadedState: {
             wallet: {
                 trading: {
                     ...tradingInitialState,
-                    sell: { ...sellInitialState, quotes: QUOTES },
+                    buy: { ...buyInitialState, quotes: QUOTES },
                 },
             },
         },
     });
 
     return renderHookWithStoreProvider(
-        ({ currentNetwork }) => {
-            const methods = useForm<TradingSellFormProps>({
+        () => {
+            const methods = useForm<TradingBuyFormProps>({
                 mode: 'onChange',
                 defaultValues,
                 resolver,
             });
-            useSellQuotes({
-                methods,
-                network: currentNetwork,
-                shouldSendInSats: false,
-                composeRequestCallback: jest.fn(),
-            });
+            useBuyQuotes({ methods, network: getNetwork('btc'), shouldSendInSats: false });
 
             return methods;
         },
-        { store, initialProps },
+        { store },
     );
 };
 
-describe('useSellQuotes', () => {
+describe('useBuyQuotes', () => {
     beforeEach(() => {
         mockHandleRequest.mockClear();
         mockAbort.mockClear();
-        mockClearQuotes.mockClear();
     });
 
     it('dispatches a quotes request after the debounce when the form is valid', async () => {
-        const { result } = renderSellQuotes(VALID_DEFAULTS);
+        const { result } = renderBuyQuotes(VALID_DEFAULTS);
 
         await act(async () => {
             await result.current.trigger();
@@ -183,6 +126,7 @@ describe('useSellQuotes', () => {
 
         expect(mockHandleRequest).toHaveBeenCalledWith(
             expect.objectContaining({
+                formValues: expect.objectContaining({ fiatInput: '100' }),
                 network: expect.objectContaining({ symbol: 'btc' }),
                 shouldSendInSats: false,
             }),
@@ -190,7 +134,7 @@ describe('useSellQuotes', () => {
     });
 
     it('auto-selects the best quote payment method when none is selected', async () => {
-        const { result } = renderSellQuotes(VALID_DEFAULTS);
+        const { result } = renderBuyQuotes(VALID_DEFAULTS);
 
         await act(async () => {
             await result.current.trigger();
@@ -199,14 +143,14 @@ describe('useSellQuotes', () => {
         await waitFor(() => expect(mockHandleRequest).toHaveBeenCalledTimes(1), { timeout: 1500 });
         await waitFor(() =>
             expect(result.current.getValues('paymentMethod')).toEqual({
-                value: 'bankTransfer',
-                label: 'Bank transfer',
+                value: 'creditCard',
+                label: 'Credit card',
             }),
         );
     });
 
     it('refetches immediately (without the debounce) when a select field changes', async () => {
-        const { result } = renderSellQuotes(VALID_DEFAULTS);
+        const { result } = renderBuyQuotes(VALID_DEFAULTS);
 
         await act(async () => {
             await result.current.trigger();
@@ -214,15 +158,15 @@ describe('useSellQuotes', () => {
         await waitFor(() => expect(mockHandleRequest).toHaveBeenCalledTimes(1), { timeout: 1500 });
 
         await act(async () => {
-            result.current.setValue('outputs.0.currency', { value: 'eur', label: 'EUR' });
+            result.current.setValue('currencySelect', { value: 'usd', label: 'USD' });
             await result.current.trigger();
         });
 
         await waitFor(() => expect(mockHandleRequest).toHaveBeenCalledTimes(2), { timeout: 200 });
     });
 
-    it('does not refetch on an output-fiat edit but does on the synced output-amount edit', async () => {
-        const { result } = renderSellQuotes(VALID_DEFAULTS);
+    it('debounces the refetch when an amount field changes', async () => {
+        const { result } = renderBuyQuotes(VALID_DEFAULTS);
 
         await act(async () => {
             await result.current.trigger();
@@ -230,21 +174,18 @@ describe('useSellQuotes', () => {
         await waitFor(() => expect(mockHandleRequest).toHaveBeenCalledTimes(1), { timeout: 1500 });
 
         await act(async () => {
-            result.current.setValue('outputs.0.fiat', '99');
+            result.current.setValue('fiatInput', '200');
             await result.current.trigger();
         });
-        await wait(NO_REFETCH_WAIT_MS);
+
+        await wait(150);
         expect(mockHandleRequest).toHaveBeenCalledTimes(1);
 
-        await act(async () => {
-            result.current.setValue('outputs.0.amount', '0.003');
-            await result.current.trigger();
-        });
         await waitFor(() => expect(mockHandleRequest).toHaveBeenCalledTimes(2), { timeout: 1500 });
     });
 
     it('does not refetch when only a non-key field (provider) changes', async () => {
-        const { result } = renderSellQuotes(VALID_DEFAULTS);
+        const { result } = renderBuyQuotes(VALID_DEFAULTS);
 
         await act(async () => {
             await result.current.trigger();
@@ -261,11 +202,11 @@ describe('useSellQuotes', () => {
     });
 
     it('does not fetch while the form is invalid', async () => {
-        const invalidResolver: Resolver<TradingSellFormProps> = () => ({
+        const invalidResolver: Resolver<TradingBuyFormProps> = () => ({
             values: {},
-            errors: { feePerUnit: { type: 'manual', message: 'invalid' } },
+            errors: { fiatInput: { type: 'manual', message: 'invalid' } },
         });
-        const { result } = renderSellQuotes(VALID_DEFAULTS, { resolver: invalidResolver });
+        const { result } = renderBuyQuotes(VALID_DEFAULTS, { resolver: invalidResolver });
 
         await act(async () => {
             await result.current.trigger();
@@ -273,16 +214,5 @@ describe('useSellQuotes', () => {
         await wait(NO_REFETCH_WAIT_MS);
 
         expect(mockHandleRequest).not.toHaveBeenCalled();
-    });
-
-    it('clears quotes eagerly when the network becomes undefined', async () => {
-        const { rerender } = renderSellQuotes(VALID_DEFAULTS);
-
-        await waitFor(() => expect(mockHandleRequest).toHaveBeenCalledTimes(1), { timeout: 1500 });
-        expect(mockClearQuotes).not.toHaveBeenCalled();
-
-        rerender({ currentNetwork: undefined });
-
-        await waitFor(() => expect(mockClearQuotes).toHaveBeenCalled());
     });
 });
